swarm_sim/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
numba>=0.57.0
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.0
pylint>=2.17.0
python-dotenv>=1.1.0 